server/requirements.txt
fastapi~=0.79.1
email-validator~=1.2.1
pyjwt~=2.4.0
psycopg2-binary~=2.9.3
tenacity~=8.0.1
cachetools~=5.2.0
more-itertools~=8.14.0
pytz~=2022.2.1